pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "scfd_vnext"
version = "0.0.1"
description = "Symplectic coherence-field dynamics vs emergent models comparison harness"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "Looptronics Lab" }]
dependencies = [
  "numpy>=1.24",
  "scipy>=1.10",
  "matplotlib>=3.7",
  "pyyaml>=6.0",
  "tqdm>=4.65",
  "torch>=2.1; platform_system!='Darwin' or platform_machine!='arm64'",
  "scikit-image>=0.21",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.3",
  "pytest-cov>=4.1",
  "pytest-xdist>=3.3",
  "pytest-rerunfailures>=11.1",
  "black>=24.3",
  "ruff>=0.5",
  "isort>=5.13",
  "mypy>=1.6",
]

[tool.setuptools]
packages = ["engine", "observe", "utils", "em_baseline", "run"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 100

[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["engine", "observe", "run"]
python_files = ["test_*.py", "tests_*.py"]
addopts = "-q"
markers = [
  "slow: long-running regression tests (energy drift, scheduler KS, predictability horizon)",
]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "UP"]

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true
warn_unused_configs = true
